Add preliminary support for generating pbs scripts
The idea is that the pbs scripts will be the interface with HPC systems,
allowing for the queueing of large numbers of jobs on the batch system
these systems have. This interface means that there no need to conduct
bash wizardry to run a batch job with a set of variables, everything is
handled in python.