Add ability to create pbs files
This allows for the creation of pbs files for running on clusters. This
is created in a very similar way to the commands in that there is a
template file with values filled in as necessary.

While this commit adds support and testing for the creation of the
files, this doesn't yet include actually running the files.