Modify processing to support pbs job submission
The processing of pbs files has a few differences to that of regular
shell commands, requiring knowledge of the jobid of the previous
command, running a set of commands at once in a batch job and also
writing the resulting file to disk.

This commit is a series of changes to the processing to support pbs job
submission and to also allow local testing of the scripts where there is
no pbs support.