Fixing type annotations to reflect compatibility
I have modified the type hints to more accurately describe the
compatibility of types and to also fix a couple of issues with
generality. This has created a PathLike type which is the union of the
str and Path types (there doesn't appear to be something already built
in that works). Also modifying input generators to be Iterables and
ChainMaps to be Mappings.