Setup versioning using version.py and bumpversion
The versioning configuration I have set up is to use a version.py file
in the module. This is loaded into the setup.py file via a function to
remove the requirement of importing the module before the module is
installed. To do the version incrementing I am using the bumpversion
tool which will automatically modify the files required and perform the
commits and tagging of version numbers.

I feel this is the simplest solution to the versioning, since I will
only have to specify what type of version is being incremented, major,
minor, or patch and bumpversion does the rest.