Change spacing around default parameters
In PEP8, there is a mention of parameters that also have type
annotations with a note that they should have spaces surrounding the
equals sign. This is only for default parameters that also have type
annotations. I was previously unaware of this convention, which is
picked up by pylint but I had assumed it was a bug.