[#13] Support for python 3.4
Python 3.6 is a very new version of python, especially in scientific
circles. This commit adds support for python 3.4, removing the use of
the new features of recent python versions.

It turns out there are a number of changes that have been made since
3.4, most notably that paths need to be explicitly converted to strings
to be useful anywhere. Additionally the updated dictionaries in python
3.6 are deterministically ordered, in previous versions the order
changes between runs which makes the testing difficult. The final change
that was required for compatibility was modifying the subprocess
commands which have been updated in python 3.5.