Create file for each iter test case
Rather than having all the tests of iteration in a single file, I have
split them into their own files to be read in and parsed by the yaml
parser. This makes it easier to edit the test cases and additionally
makes them a better example of the possible ways of combining variables.