Breaking Change - Modify behaviour of iterators
I have modified how the iterators work with the dictionary, rather than
being a value in the dictionary determining the behaviour, I feel it
makes more sense to have them defining the way the contained variables
interact.

In the old version the following would have been valid
```
var1:
    value: [1, 2, 3]
    var2: [2, 4, 6]
    iterator: zip
```
It seems unusual to have a different format for var1 over var2. So I
have removed the need to the value, instead using the iterator as the
major dictionary
```
zip:
    var1: [1, 2, 3]
    var2: [2, 4, 6]
```
which is hopefully a lot easier to understand and work with.