Allow arbitrary pbs specifications
I have expanded the parsing capability of the pbs options, with a major
feature being the inclusion of arbitrary options. This is done in a way
that passing a flag without the dash `-` as the key will create the
appropriate entry. This allows for complex specifications like
```
pbs:
    M: malramsay64@gmail.com
    j: oe
```
There is still the issue that this is not particularly use friendly,
however it remains possible.

In addition the parsing of the standard options is better allowing for
the specification of project, memory, and gpus.