Bugfix - dictionary modified instead of copied
There was a bug where a value was been added to a dictionary
unnecessarily, however this was only causing a problem as it was
updating the original dictionary. This commit removes both these issues
by removing the offending line of code, and by copying the resulting
dictionary when working on the values contained within.

Additionally this means that this now works on artemis completing #7