Allow for jinja2 templates in environment.yml
Sometimes you are too lazy to type out all the values of the variables,
or there are just too many of them. Sometimes you want to specify the
variables programmatically. Using jinja2 templates allows for a much
more expressive way of defining variables.

While jinja2 isn't the most intuitive format, there are a lot of other
projects using it, and most importantly there is basically no code for
me to write which means this can be implemented super quickly.