MAINT: Update type hinting to remove errors
There were a few errors in the type hints which I hadn't noticed
previously. This fixes some of the issues. There are still a few
locations where the type hints are not being helpful, primarily the zip
and product iterators, which is indicative of me not knowing how to
handle the typing of them.