TST Fuzz testing of variable names
This introduces testing for valid variable names for Experi. Currently I
am only testing my python logic, i.e. not parsing the yaml input file.
The test I am running is for at least a single unicode letter character,
followed by any unicode character which isn't one of [":", "{", "}", ".",
"!"]. This is not exhaustive testing and there is possibly going to be
some random errors which pop up, although this is reassuring as to the
generality of the code.