Create function which generates a conditional bash command
Currently I think the most sensible place to put the logic for
conditionally running a command is in the command to be run itself. This
means that the jobs can be submitted from the generated pbs files and
only run the commands that are required. While this does mean more jobs
sitting in the queue for a while, it is possibly better than the
generated files changing regularly.