ENH Provide list of values to afterok qsub option
Adding a list of values is the documented method of giving a job
multiple dependencies. The idea is that all subsequent jobs will also
fail when a prior job fails. It is possible that different pbs versions
are handling multiple -W options differently hence the different
behaviour between clusters. This fixes #27.