ENH Incorporate scheduler options into Job class
By including the scheduler options in the job class this paves the way
for job specific options, like the number cores for a specific job. In
making this change I have also made the process of integrating an
additional scheduler much simpler by having the logic split in it's own
function.