ENH Command class supports a list of commands
The command class allows for the running of a list of commands by
specifying the list as the value of the command key. With this change
there is now the ability to specify whether a line of shell code is submitted as
as discrete jobs using the jobs key, or as a line of a command within a
job.

This fixes #30, which makes tasks like making a directory before running
a command somewhat simpler as it can take place within the same job
submission script.