ENH Replace combine with append and cycle iterators
The implementation of the combine iterator wasn't the iterator I thought
I was creating and basically it can already be achieved better with the
zip iterator. Instead I have replaced it with the append iterator which
does what I was initially intending and the cycle iterator which cycles
through values.