ENH Support skipping simulations when files exist
This adds the `--use-dependencies` flag, which when used will only run
the simulations for which the file specified in `creates` of a command
does not yet exist. The default is to run all the simulations, with this
being a choice. This only checks for the presence of the file with no
regards for modification time or having some contents.