ENH Allow specification of scheduler on the command line
This makes a number of changes to allow for the scheduler to be
specified on the command line, and when using the launch command. This
functionality has been added for a couple of reasons, the obvious new
adding features reason, however also to make it significantly easier
to test the functionality of both the PBS and SLURM schedulers. By
enabling the scheduler to be specified separately to the file I am able
to use the same files to test the generation of files of both schedulers
and more importantly check they both run to some degree.